refactor: split markdown.py into table utilities and eval parsing modules
